test_requirements.txt
bumpversion>=0.5.3
django-test-plus
wheel>=0.30.0
coverage>=4.4.1
mock>=1.0.1
flake8>=2.1.0
codecov>=2.0.0
pytest-django
pytest-cov
pytest-sugar
pytest-xdist
pytest
pipenv
pytest-flake8
sphinx
recommonmark
sphinx-rtd-theme


